rename dialog_validation to dialog_form
