+++ app/javascript/src/component_dialog_form.js
