+++ test/dialogs/dialog_form/component_test.js
