+++ test/dialogs/dialog_form/methods_test.js
