+++ test/dialogs/dialog_form/properties_test.js
